requirements.txt
pillow==5.3.0
termcolor==1.1.0